Remove model.py
Replaced with more appropriately named `agent.py` and `message.py`.
