+++ python/agent.py
@@ -1,0 +1,15 @@
+""" Module defining Agent class and related methods
+"""
+class Agent:
+    """ Agent class storing all needed elements for agent operation.
+    """
+    def __init__(self):
+        self.owner = None
+        self.wallet_handle = None
+        self.endpoint = None
+        self.endpoint_vk = None
+        self.ui_token = None
+        self.pool_handle = None
+        self.ui_socket = None
+        self.initialized = False
+        self.modules = []
